地址类

+++ rocket/net/io_thread_group.h
@@ -15,7 +15,7 @@ namespace rocket
     private:
         int m_size{0};
         std::vector<IOThread*> m_io_thread_groups;
-        int m_index {0};
+        size_t m_index {0};
 
     public:
         IOThreadGroup(int size);
